Final Excel + Bug Fixes + Validations

+++ KanaksTiffins/KanakTiffins/PopulateMaster.cs
@@ -44,11 +44,18 @@ namespace KanakTiffins
            if (textBox_editArea.Text.Trim().Length  != 0)
            {
                String areaName = comboBox_areas.SelectedValue.ToString();
+               if (db.Areas.Select(x => x.AreaName).Contains(textBox_editArea.Text))
+               {
+                   MessageBox.Show("Already Exists.", "Error");
+                   return;
+               }
                db.Areas.Where(x => x.AreaName == areaName).Single().AreaName = textBox_editArea.Text;
                db.SaveChanges();
                //Refersh the Area combo-box.
                CommonUtilities.populateAreas(comboBox_areas);
-               MessageBox.Show("Updated Successfully", "Success");               
+               MessageBox.Show("Updated Successfully", "Success");
+               textBox_editArea.Text = "";
+               return;
            }
 
            if (Int32.TryParse(textBox_area.Text.Trim(), out textValue))
@@ -64,7 +71,8 @@ namespace KanakTiffins
             }
 
             //Validation succeeded.
-
+            if (textBox_area.Text.Trim().Length != 0)
+            {
                 int lastAreaId = 0;
 
                 if (db.Areas.Count() != 0)
@@ -79,6 +87,8 @@ namespace KanakTiffins
                 db.SaveChanges();
                 //Refersh the Area combo-box.
                 CommonUtilities.populateAreas(comboBox_areas);
+                textBox_area.Text = "";
+            }
         }
 
         /// <summary>
@@ -95,6 +105,23 @@ namespace KanakTiffins
                 return;
             }
 
+            if (textBox_editMealPlan.Text.Trim().Length != 0)
+            {
+                MealPlan mealPlan = comboBox_mealPlans.SelectedItem as MealPlan;
+
+                if (db.MealPlans.Select(x => x.MealAmount).Contains(Int32.Parse(textBox_editMealPlan.Text)))
+                {
+                    MessageBox.Show("Already Exists.", "Error");
+                    return;                
+                }
+                db.MealPlans.Where(x => x.MealAmount == mealPlan.MealAmount).Single().MealAmount = Int32.Parse(textBox_editMealPlan.Text);
+                db.SaveChanges();
+                //Refresh the Area combo-box.
+                CommonUtilities.populateMealPlans(comboBox_mealPlans);
+                MessageBox.Show("Updated Successfully", "Success");
+                textBox_editMealPlan.Text = "";
+                return;
+            }
 
             int textValue = 0;
             if (!Int32.TryParse(textBox_mealPlan.Text.Trim(), out textValue))
@@ -104,7 +131,7 @@ namespace KanakTiffins
             }
             if (Int32.Parse(textBox_mealPlan.Text) <= 0)
             {
-                MessageBox.Show("Cant Be a Negative Number.", "Error");
+                MessageBox.Show("Cant Be Zero or A Negative Number.", "Error");
                 return;
             }
             if (db.MealPlans.Select(x => x.MealAmount).Contains(Int32.Parse(textBox_mealPlan.Text)))
@@ -115,23 +142,32 @@ namespace KanakTiffins
             
             //Validation succeeded.
 
-            int lastMealPlanId = db.MealPlans.Select(x => x.MealPlanId).Max();
-            MealPlan newMealPlan = new MealPlan();
-            newMealPlan.MealPlanId = lastMealPlanId + 1;
-            newMealPlan.MealAmount = textBox_mealPlan.Text == "" ? 0 : Int32.Parse(textBox_mealPlan.Text);
 
-            if (newMealPlan.MealAmount == 0)
+            if (textBox_mealPlan.Text.Trim().Length != 0)
             {
-                MessageBox.Show("Cannot be 0", "Error");
-                return;
+                int lastMealPlanId = 0;
+
+                if (db.MealPlans.Count() != 0)
+                    lastMealPlanId = db.MealPlans.Select(x => x.MealPlanId).Max();
+
+                MealPlan newMealPlan = new MealPlan();
+                newMealPlan.MealPlanId = lastMealPlanId + 1;
+                newMealPlan.MealAmount = textBox_mealPlan.Text == "" ? 0 : Int32.Parse(textBox_mealPlan.Text);
+
+                if (newMealPlan.MealAmount == 0)
+                {
+                    MessageBox.Show("Cannot be 0", "Error");
+                    return;
+                }
+                db.MealPlans.AddObject(newMealPlan);
+
+                db.SaveChanges();
+                MessageBox.Show("Added Successfully", "Success");
+
+                //Refresh the MealPlan combo-box.
+                CommonUtilities.populateMealPlans(comboBox_mealPlans);
+                textBox_mealPlan.Text = "";
             }
-            db.MealPlans.AddObject(newMealPlan);
-
-            db.SaveChanges();
-            MessageBox.Show("Added Successfully", "Success");
-
-            //Refresh the MealPlan combo-box.
-            CommonUtilities.populateMealPlans(comboBox_mealPlans);
         }
 
         /// <summary>
@@ -143,6 +179,7 @@ namespace KanakTiffins
         {
             Area selectedArea = comboBox_areas.SelectedItem as Area;
 
+            //To check if a proper value has been selected
             if (selectedArea.AreaName != null)
             {
                 //Check whether this AreaId has already been used anywhere.
@@ -156,7 +193,7 @@ namespace KanakTiffins
                 db.Areas.DeleteObject(selectedArea);
                 db.SaveChanges();
                 CommonUtilities.populateAreas(comboBox_areas);
-                MessageBox.Show("Deleted Successfully", "Success");
+                MessageBox.Show("Deleted Successfully", "Success");                
             }
             else
             {
@@ -173,22 +210,32 @@ namespace KanakTiffins
         private void button_deleteMealPlan_Click(object sender, EventArgs e)
         {
             MealPlan selectedMealPlan = comboBox_mealPlans.SelectedItem as MealPlan;
-
-            //Check whether this MealPlanId has already been used anywhere.
-            if (db.CustomerDetails.Select(x => x.MealPlanId).Contains(selectedMealPlan.MealPlanId))
+           
+            //To check if a proper value has been selected
+            if (selectedMealPlan != null)
             {
-                MessageBox.Show("Cannot delete this Meal Plan (" + selectedMealPlan.MealAmount + "). It is already being used by one/many customers. Please contact the developers for more information.", "Error");
+                //Check whether this MealPlanId has already been used anywhere.
+                if (db.CustomerDetails.Select(x => x.MealPlanId).Contains(selectedMealPlan.MealPlanId))
+                {
+                    MessageBox.Show("Cannot delete this Meal Plan (" + selectedMealPlan.MealAmount + "). It is already being used by one/many customers. Please contact the developers for more information.", "Error");
+                    return;
+                }
+
+                //This Meal Plan can be deleted.
+                db.MealPlans.DeleteObject(selectedMealPlan);
+                db.SaveChanges();
+                CommonUtilities.populateMealPlans(comboBox_mealPlans);
+                MessageBox.Show("Deleted Successfully", "Success");
+                comboBox_mealPlans.Text = "";                
+            }
+            else
+            {
+                MessageBox.Show("Please Select the Meal Plan to be Deleted", "Error");
                 return;
             }
-
-            //This Meal Plan can be deleted.
-            db.MealPlans.DeleteObject(selectedMealPlan);
-            db.SaveChanges();
-            CommonUtilities.populateMealPlans(comboBox_mealPlans);
-            MessageBox.Show("Deleted Successfully", "Success");
         }
         /// <summary>
-        /// On Change of an Area, rendering appropriate value on the textboxes
+        /// On Change of an Area, rendering appropriate value in the textboxes
         /// </summary>
         /// <param name="sender"></param>
         /// <param name="e"></param>
